+++ pycricinfo/config.py
@@ -9,6 +9,7 @@ class BaseRoute(Enum):
     core: str = auto()
     site: str = auto()
     page: str = auto()
+    stats: str = auto()
 
 
 class CoreAPIRoutes(BaseModel):
@@ -43,6 +44,7 @@ class PageRoutes(BaseModel):
     player_profile: str = "cricketers/{player_slug}-{player_id}"
     series_in_season: str = "ci/engine/series/index.html?season={season_name};view=season"
     matches_in_series: str = "ci/engine/match/index/series.html?series={series_id}"
+    player_stats: str = "ci/engine/player/{player_id}.html?class=11;template=results;type={stat_type}"
 
 
 class PageHeaders(BaseModel):
@@ -54,6 +56,7 @@ class Settings(BaseSettings):
     core_base_route_v2: str = "http://core.espnuk.org/v2/sports/cricket/"
     site_base_route_v2: str = "http://site.api.espn.com/apis/site/v2/sports/cricket/"
     cricinfo_base_route: str = "https://www.espncricinfo.com/"
+    stats_base_route: str = "https://stats.espncricinfo.com/"
 
     routes: CoreAPIRoutes = CoreAPIRoutes()
     page_routes: PageRoutes = PageRoutes()