+++ changelog.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.0.37]
-### Fixed
-- Help avoid 403s when scraping pages
+### Changed
+- Move to collecting Player career data from Statsguru instead of page scraping, due to bot detection
+
+### Removed
+- The method of fetching Player career data added in `[0.0.35]`
 
 ## [0.0.36]
 ### Fixed