+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.38]
+### Fixed
+- Handle Career stats for Players who have only played one format
+
 ## [0.0.37]
 ### Changed
 - Move to collecting Player career data from Statsguru instead of page scraping, due to bot detection